requirements.txt
allure-behave==2.13.5
allure-python-commons==2.13.5
behave==1.2.6
playwright==1.53.0
python-dotenv==1.0.1
colorama~=0.4.6